added single line editor with highlighting
