--- app/src/lib/design-system-v2/components/RQInput/RQInput.scss
@@ -1,52 +1,0 @@
-/* Editor container styling */
-// .single-line-editor-container {
-//   width: 100%;
-//   box-sizing: border-box;
-
-//   .CodeMirror {
-//     height: auto !important;
-//     white-space: nowrap !important; /* Prevent line wrapping */
-//     overflow-x: hidden; /* Allow horizontal scrolling */
-//     caret-color: white !important; /* Make sure caret is visible */
-//     background-color: transparent; /* Dark background */
-//     color: inherit !important; /* Light text */
-//     box-sizing: border-box;
-
-//     .cm-env-variable {
-//       color: red; /* Make the environment variables red */
-//     }
-
-//     .CodeMirror-cursor {
-//       border-left: 1px solid white;
-//     }
-
-//     div.CodeMirror-selected,
-//     .CodeMirror span.CodeMirror-selectedtext {
-//       background-color: highlight !important; /* Uses the browser's default highlight color */
-//       color: highlighttext !important; /* Ensure text color is the default highlight text color */
-//     }
-//   }
-// }
-
-.single-line-editor-container {
-  .highlight-variable {
-    color: red;
-    cursor: pointer;
-  }
-
-  .cm-focused {
-    outline: none;
-  }
-}
-
-div.n-popup {
-  position: absolute;
-  background-color: #333;
-  color: white;
-  padding: 5px;
-  border-radius: 3px;
-  white-space: nowrap;
-  font-size: 12px;
-  z-index: 10;
-  pointer-events: none; /* Prevent interaction with the popup */
-}
