+++ app/src/lib/design-system-v2/components/SingleLineEditor/SingleLineEditor.scss
@@ -1,0 +1,22 @@
+.single-line-editor-container {
+  .highlight-variable {
+    color: red;
+    cursor: pointer;
+  }
+
+  .cm-focused {
+    outline: none;
+  }
+}
+
+div.n-popup {
+  position: absolute;
+  background-color: #333;
+  color: white;
+  padding: 5px;
+  border-radius: 3px;
+  white-space: nowrap;
+  font-size: 12px;
+  z-index: 10;
+  pointer-events: none; /* Prevent interaction with the popup */
+}
