+++ app/src/lib/design-system-v2/components/index.ts
@@ -1,2 +1,2 @@
 export * from "./RQButton/RQButton";
-export * from "./RQInput/RQInput";
+export * from "./SingleLineEditor/SingleLineEditor";
